restaurant_management_api/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-Session==0.5.0
python-dotenv==1.0.0
SQLAlchemy==2.0.21
Werkzeug==2.3.7
alembic==1.12.0
bcrypt==4.0.1
click==8.1.7
colorama==0.4.6
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
typing_extensions==4.8.0
python-dateutil==2.8.2
six==1.16.0
requests==2.31.0

# Security and Validation Dependencies
marshmallow==3.20.1
flask-limiter==3.5.0

# AI/ML Libraries
scikit-learn==1.3.0
pandas==2.1.1
numpy==1.25.2
matplotlib==3.7.2
seaborn==0.12.2

# Time Series Analysis
statsmodels==0.14.0
# prophet==1.1.4

# NLP
spacy==3.6.0
nltk==3.8.1

# Computer Vision
opencv-python==4.8.0.76
pytesseract==0.3.10
Pillow==10.0.0

# Model Management
mlflow==2.6.0
joblib==1.3.2

# Enhanced Backend
fastapi==0.103.1
uvicorn==0.23.2
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.7

# Monitoring
prometheus-client==0.17.1

# Compatibility Fix
watchdog==1.0.2

# New dependencies
plotly==5.17.0
openpyxl==3.1.2
xlrd==2.0.1
gunicorn==21.2.0
